+++ src/components/vertical-menu/responsive-vertical-menu/responsive-vertical-menu.pw.tsx
@@ -7,6 +7,7 @@ import {
   WithDifferentDepthsAsLastItem,
   WithLongList,
   WithButtonAndLinkActionItems,
+  WithSiblingControl,
 } from "./components.test-pw";
 
 import {
@@ -222,6 +223,33 @@ test.describe("when responsive vertical menu item renders as an anchor", () => {
 
     await expect(responsiveVerticalMenuPrimary(page)).toBeVisible();
   });
+
+  test("background controls cannot be interacted with while menu is open", async ({
+    mount,
+    page,
+  }) => {
+    await mount(<WithSiblingControl />);
+
+    const launcherButton = page.getByRole("button", {
+      name: /Product menu launcher/,
+    });
+    await launcherButton.click();
+
+    const menuItem = page.getByRole("button", {
+      name: /Primary Menu With Children/,
+    });
+    await menuItem.waitFor();
+
+    const outsideCheckbox = page.getByRole("checkbox");
+
+    const { x, y } = (await outsideCheckbox.boundingBox()) as {
+      x: number;
+      y: number;
+    };
+    await page.mouse.click(x, y);
+
+    await expect(outsideCheckbox).not.toBeChecked();
+  });
 });
 
 test.describe("when responsive vertical menu item renders as a button", () => {
@@ -297,7 +325,7 @@ test.describe("keyboard navigation tests", () => {
     await expect(firstMenuItem).toBeInViewport();
   });
 
-  test("should loop back to primary menu item when Tab is pressed on last secondary menu item", async ({
+  test("should navigate to the next parent item when last menu item is depth 2", async ({
     mount,
     page,
   }) => {
@@ -317,10 +345,14 @@ test.describe("keyboard navigation tests", () => {
     await page.keyboard.press("Tab");
     await expect(secondaryMenuItem).toBeFocused();
     await page.keyboard.press("Tab");
-    await expect(primaryMenuItem).toBeFocused();
+
+    const nextPrimaryMenuItem = page.getByRole("button", {
+      name: "With Level 3",
+    });
+    await expect(nextPrimaryMenuItem).toBeFocused();
   });
 
-  test("should loop back to primary menu item when Tab is pressed on last tertiary menu item", async ({
+  test("should navigate to the next parent item when last menu item is depth 3", async ({
     mount,
     page,
   }) => {
@@ -344,10 +376,14 @@ test.describe("keyboard navigation tests", () => {
     await page.keyboard.press("Tab");
     await expect(tertiaryMenuItem).toBeFocused();
     await page.keyboard.press("Tab");
-    await expect(primaryMenuItem).toBeFocused();
+
+    const nextPrimaryMenuItem = page.getByRole("button", {
+      name: "With Level 4",
+    });
+    await expect(nextPrimaryMenuItem).toBeFocused();
   });
 
-  test("should loop back to primary menu item when Tab is pressed on last level 4 item", async ({
+  test("should navigate to the next parent item when last menu item is depth 4", async ({
     mount,
     page,
   }) => {
@@ -374,11 +410,32 @@ test.describe("keyboard navigation tests", () => {
     await expect(quaternaryMenuItem).toBeFocused();
     await page.keyboard.press("Tab");
 
-    // TODO: FE-7448
-    // await expect(primaryMenuItem).toBeFocused();
+    const nextPrimaryMenuItem = page.getByRole("button", {
+      name: "Last Primary Menu Item",
+    });
+    await expect(nextPrimaryMenuItem).toBeFocused();
   });
 
-  test("should go to last secondary menu item when Shift+Tab is pressed on the primary menu", async ({
+  test("should close the menu and focus the next element in the tab order when tabbing out of the menu", async ({
+    mount,
+    page,
+  }) => {
+    await mount(<WithSiblingControl />);
+
+    await responsiveVerticalMenuLauncher(page).click();
+
+    const lastItem = page.getByRole("button", { name: "Primary Menu Item" });
+    await page.keyboard.press("Tab");
+    await page.keyboard.press("Tab");
+    await expect(lastItem).toBeFocused();
+
+    // press tab to move focus out of the menu
+    await page.keyboard.press("Tab");
+    const externalInput = page.getByRole("checkbox");
+    await expect(externalInput).toBeFocused();
+  });
+
+  test("should go to last secondary menu item when Shift+Tab is pressed on the next parent item", async ({
     mount,
     page,
   }) => {
@@ -387,21 +444,25 @@ test.describe("keyboard navigation tests", () => {
     await responsiveVerticalMenuLauncher(page).click();
 
     const primaryMenuItem = page.getByRole("button", { name: "With Level 2" });
-    await page.keyboard.press("Tab");
-    await expect(primaryMenuItem).toBeFocused();
-    await page.keyboard.press("Enter");
+    const nextPrimaryMenuItem = page.getByRole("button", {
+      name: "With Level 3",
+    });
 
+    await primaryMenuItem.click(); // open primary menu
     const secondaryMenuItem = page.getByRole("link", {
       name: "Level 2 As Last Item",
     });
     await expect(secondaryMenuItem).toBeVisible();
+
+    await nextPrimaryMenuItem.focus();
+
     await page.keyboard.press("Shift+Tab");
     await expect(secondaryMenuItem).toBeFocused();
     await page.keyboard.press("Shift+Tab");
     await expect(primaryMenuItem).toBeFocused();
   });
 
-  test("should go to last tertiary menu item when Shift+Tab is pressed on the secondary menu", async ({
+  test("should go to last tertiary menu item when Shift+Tab is pressed on the next parent item", async ({
     mount,
     page,
   }) => {
@@ -410,6 +471,10 @@ test.describe("keyboard navigation tests", () => {
     await responsiveVerticalMenuLauncher(page).click();
 
     const primaryMenuItem = page.getByRole("button", { name: "With Level 3" });
+    const nextPrimaryMenuItem = page.getByRole("button", {
+      name: "With Level 4",
+    });
+
     await primaryMenuItem.click(); // open primary menu
     const secondaryMenuItem = page.getByRole("button", {
       name: "Level 2 Parent",
@@ -420,7 +485,7 @@ test.describe("keyboard navigation tests", () => {
     });
     await expect(tertiaryMenuItem).toBeVisible();
 
-    await primaryMenuItem.focus();
+    await nextPrimaryMenuItem.focus();
 
     await page.keyboard.press("Shift+Tab");
     await expect(tertiaryMenuItem).toBeFocused();
@@ -430,8 +495,7 @@ test.describe("keyboard navigation tests", () => {
     await expect(primaryMenuItem).toBeFocused();
   });
 
-  // TODO: FE-7448
-  test.skip("should go to last level 4 item when Shift+Tab is pressed on the tertiary menu", async ({
+  test("should go to last level 4 item when Shift+Tab is pressed on the next parent item", async ({
     mount,
     page,
   }) => {
@@ -440,6 +504,10 @@ test.describe("keyboard navigation tests", () => {
     await responsiveVerticalMenuLauncher(page).click();
 
     const primaryMenuItem = page.getByRole("button", { name: "With Level 4" });
+    const nextPrimaryMenuItem = page.getByRole("button", {
+      name: "Last Primary Menu Item",
+    });
+
     await primaryMenuItem.click(); // open primary menu
     const secondaryMenuItem = page.getByRole("button", {
       name: "Level 2 Parent",
@@ -453,7 +521,7 @@ test.describe("keyboard navigation tests", () => {
     });
     await expect(quaternaryMenuItem).toBeVisible();
 
-    await primaryMenuItem.focus();
+    await nextPrimaryMenuItem.focus();
 
     await page.keyboard.press("Shift+Tab");
     await expect(quaternaryMenuItem).toBeFocused();
@@ -539,6 +607,11 @@ test.describe("accessibility tests", () => {
     await page.locator("[id='primary-menu']").click();
     await page.locator("[id='secondary-menu']").click();
     await expect(page.locator("[id='tertiary-menu']")).toBeVisible();
+
+    // Wait for fade in animation to complete before scanning
+    // to ensure colours aren't slightly transparent
+    const modal = page.getByTestId("modal");
+    await expect(modal).toHaveAttribute("data-state", "open");
 
     await checkAccessibility(page);
   });